LayoutTests/inspector/debugger/setShouldBlackboxURL-regex-case-sensitive-expected.txt

Tests Debugger.setShouldBlackboxURL.


== Running test suite: Debugger.setShouldBlackboxURL
-- Running test case: Debugger.setShouldBlackboxURL.CaseSensitiveRegex.stepOver
Evaluating 'createScripts("CaseSensitiveRegex_StepOver")'...
Blackboxing (case sensitive) (regex) 'casesensitiveregex_stepover_inner\.js$'...
Blackboxing (case sensitive) (regex) 'CaseSensitiveRegex_StepOver_Middle\.js$'...
Setting breakpoint in 'CaseSensitiveRegex_StepOver_Inner.js'...
Evaluating 'CaseSensitiveRegex_StepOver_Outer(10)'...

PAUSED: 'Breakpoint' at 'CaseSensitiveRegex_StepOver_Inner:3:1'.
{
  "breakpointId": "CaseSensitiveRegex_StepOver_Inner.js:3:0"
}
Stepping over...

PAUSED: 'BlackboxedScript' at 'CaseSensitiveRegex_StepOver_Outer:3:1'.
{
  "originalReason": "other",
  "originalData": {
    "breakpointId": "CaseSensitiveRegex_StepOver_Inner.js:3:0"
  }
}
Stepping over...

Resuming...
PASS: Resumed.
PASS: Should pause in 'CaseSensitiveRegex_StepOver_Inner'.
PASS: Should not pause in 'CaseSensitiveRegex_StepOver_Middle'.


-- Running test case: Debugger.setShouldBlackboxURL.CaseSensitiveRegex.PauseInCaller
Evaluating 'createScripts("CaseSensitiveRegex_PauseInCaller")'...
Blackboxing (case sensitive) (regex) 'CaseSensitiveRegex_PauseInCaller_Inner\.js$'...
Blackboxing (case sensitive) (regex) 'casesensitiveregex_pauseincaller_middle\.js$'...
Setting breakpoint in 'CaseSensitiveRegex_PauseInCaller_Inner.js'...
Evaluating 'CaseSensitiveRegex_PauseInCaller_Outer(10)'...

PAUSED: 'BlackboxedScript' at 'CaseSensitiveRegex_PauseInCaller_Middle:3:1'.
{
  "originalReason": "Breakpoint",
  "originalData": {
    "breakpointId": "CaseSensitiveRegex_PauseInCaller_Inner.js:2:0"
  }
}
Stepping over...

PAUSED: 'other' at 'CaseSensitiveRegex_PauseInCaller_Outer:3:1'.
Stepping over...

Resuming...
PASS: Resumed.
PASS: Should not pause in 'CaseSensitiveRegex_PauseInCaller_Inner'.
PASS: Should pause in 'CaseSensitiveRegex_PauseInCaller_Middle'.


-- Running test case: Debugger.setShouldBlackboxURL.CaseSensitiveRegex.PauseInCallee
Evaluating 'createScripts("CaseSensitiveRegex_PauseInCallee")'...
Blackboxing (case sensitive) (regex) 'casesensitiveregex_pauseincallee_middle\.js$'...
Blackboxing (case sensitive) (regex) 'CaseSensitiveRegex_PauseInCallee_Outer\.js$'...
Setting breakpoint in 'CaseSensitiveRegex_PauseInCallee_Outer.js'...
Evaluating 'CaseSensitiveRegex_PauseInCallee_Outer(10)'...

PAUSED: 'BlackboxedScript' at 'CaseSensitiveRegex_PauseInCallee_Middle:2:4'.
{
  "originalReason": "Breakpoint",
  "originalData": {
    "breakpointId": "CaseSensitiveRegex_PauseInCallee_Outer.js:2:0"
  }
}
Stepping over...

PAUSED: 'other' at 'CaseSensitiveRegex_PauseInCallee_Middle:3:1'.
Stepping over...

Resuming...
PASS: Resumed.
PASS: Should pause in 'CaseSensitiveRegex_PauseInCallee_Middle'.
PASS: Should not pause in 'CaseSensitiveRegex_PauseInCallee_Outer'.


-- Running test case: Debugger.setShouldBlackboxURL.CaseSensitiveRegex.SourceRange
Evaluating 'createScripts("CaseSensitiveRegex_Range")'...
Blackboxing (case sensitive) (regex) 'casesensitiveregex_range_multiple\.js$'...
Blackboxing (case sensitive) (regex) 'CaseSensitiveRegex_Range_Multiple\.js$' 3:0-4:0...
Blackboxing (case sensitive) (regex) 'CaseSensitiveRegex_Range_Multiple\.js$' 4:5-4:18...
Blackboxing (case sensitive) (regex) 'CaseSensitiveRegex_Range_Multiple\.js$' 5:1-5:2...
Setting breakpoint in 'CaseSensitiveRegex_Range_Multiple.js'...
Evaluating 'CaseSensitiveRegex_Range_Multiple(10)'...

PAUSED: 'Breakpoint' at 'CaseSensitiveRegex_Range_Multiple:2:4'.
{
  "breakpointId": "CaseSensitiveRegex_Range_Multiple.js:2:0"
}
Stepping over...

PAUSED: 'BlackboxedScript' at 'CaseSensitiveRegex_Range_Multiple:4:4'.
{
  "originalReason": "other",
  "originalData": {
    "breakpointId": "CaseSensitiveRegex_Range_Multiple.js:2:0"
  }
}
Stepping over...

Resuming...
PASS: Resumed.
PASS: Should pause in 'CaseSensitiveRegex_Range_Multiple'.
